+++ NEWS.md
@@ -3,8 +3,8 @@ Known issues: https://github.com/PredictiveEcology/reproducible/issues
 version 1.2.6.9000
 ==============
 
-## Ne features
-* none
+## Enhancements
+* `quick` argument in `Cache` can now be a character vector, allowing individual character arguments to be digested as character vectors and others to be digested as files located at the specified path as represented by the character vector.
 
 ## Enhancements
 * none